pyproject.toml
[project]
name = "comfyui-rtx-remix"
description = "Use ComfyUI with RTX Remix to remaster classic games [a/https://github.com/NVIDIAGameWorks/rtx-remix](https://github.com/NVIDIAGameWorks/rtx-remix)"
version = "1.1.0"
license = {file = "LICENSE"}
dependencies = ["numpy", "pillow>=10.0.1", "requests", "torch"]

[project.urls]
Repository = "https://github.com/NVIDIAGameWorks/ComfyUI-RTX-Remix"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "nvida-gameworks"
DisplayName = "ComfyUI-RTX-Remix"
Icon = "images/remix.png"
